Resolve GitHub CVE advisory by updating to latest elliptic
In versions of elliptic prior to 6.5.4, there is no check to confirm that the public key point passed into the derive function actually exists on the secp256k1 curve. This results in the potential for the private key used in this implementation to be revealed after a number of ECDH operations are performed.

See https://github.com/advisories/GHSA-r9p9-mrjm-926w